requirements.txt
flask
google-cloud-storage
google-cloud-bigquery
gunicorn
